Move store tests into their own folder
