+++ app/test/unit/stores/git-store-test.ts
@@ -5,18 +5,18 @@ import * as Fs from 'fs'
 import * as Path from 'path'
 import { GitProcess } from 'dugite'
 
-import { shell } from '../helpers/test-app-shell'
+import { shell } from '../../helpers/test-app-shell'
 import {
   setupEmptyRepository,
   setupFixtureRepository,
   setupConflictedRepo,
-} from '../helpers/repositories'
-import { GitStore } from '../../src/lib/stores'
-import { AppFileStatus } from '../../src/models/status'
-import { Repository } from '../../src/models/repository'
-import { Commit } from '../../src/models/commit'
-import { TipState, IValidBranch } from '../../src/models/tip'
-import { getCommit, getStatus } from '../../src/lib/git'
+} from '../../helpers/repositories'
+import { GitStore } from '../../../src/lib/stores'
+import { AppFileStatus } from '../../../src/models/status'
+import { Repository } from '../../../src/models/repository'
+import { Commit } from '../../../src/models/commit'
+import { TipState, IValidBranch } from '../../../src/models/tip'
+import { getCommit, getStatus } from '../../../src/lib/git'
 
 describe('GitStore', () => {
   it('can discard changes from a repository', async () => {
